new version snake | add new file [Action | Setting] | edit code| add menu
